requirements.txt
Django==1.8.5
Werkzeug==0.10.4
argparse==1.2.1
django-extensions==1.5.7
djangorestframework==3.2.4
psycopg2==2.6.1
six==1.10.0
wsgiref==0.1.2
